- genai summary support - refactoring and modularization - requirements upgrade - documentation
